(BSR)[API] refactor: Move common incident methods in api file
